Split up ty tests a bit
